[kernel][x86] Allow modifying 3 more flag bits via mx_thread_write_state()
Before this change, using mx_thread_write_state() would always reset
the NT, AC and ID bits to zero and ignore the values passed in for
those bits.  These three are safe to allow modifying because they can
be modified from userland.

Of these three, AC is the only one that's useful to set, but we should
allow all three for completeness and to ensure that round-tripping the
register state via mx_thread_read_state()+mx_thread_write_state()
doesn't modify the register state.

Also allow reading and writing these flags via the hypervisor
syscalls, mx_vcpu_read_state() and mx_vcpu_write_state().

+++ system/public/magenta/syscalls/hypervisor.h
@@ -97,8 +97,8 @@ typedef struct mx_vcpu_state {
     uint64_t r13;
     uint64_t r14;
     uint64_t r15;
-    // Only the user-controllable lower 16-bits of the flags register.
-    uint16_t flags;
+    // Only the user-controllable lower 32-bits of the flags register.
+    uint32_t flags;
 #endif
 } mx_vcpu_state_t;
